refactor: 代码移动，删除加载音乐的代码
